+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add support for `FunctionTemplate.Inherit` to set up prototype inheritance.
+
 ### Changed
 
 ## [v0.32.0] - 2025-04-30